clang-tidy: default to define default contructor; redundant boolean literal

+++ src/bd_fit.cpp
@@ -91,9 +91,9 @@ void BDFit::fit_yule () {
 
 
 void BDFit::fit_bd() {
-    branching_times_.resize(nintnodes_);
-    for (int i = 0; i < nintnodes_; i++) {
-        branching_times_[static_cast<unsigned long>(i)] = tree_->getInternalNode(i)->getHeight();
+    branching_times_.resize(static_cast<unsigned int>(nintnodes_));
+    for (unsigned int i = 0; i < static_cast<unsigned int>(nintnodes_); i++) {
+        branching_times_[i] = tree_->getInternalNode(i)->getHeight();
     }
     
     // sort in descending order
